+++ core/src/main/java/net/nevinsky/abyssus/core/builder/CylinderShapeBuilder.java
@@ -60,7 +60,7 @@ public class CylinderShapeBuilder extends BaseShapeBuilder {
         curr1.hasUV = curr1.hasPosition = curr1.hasNormal = true;
         MeshPartBuilder.VertexInfo curr2 = vertTmp4.set(null, null, null, null);
         curr2.hasUV = curr2.hasPosition = curr2.hasNormal = true;
-        short i1, i2, i3 = 0, i4 = 0;
+        int i1, i2, i3 = 0, i4 = 0;
 
         builder.ensureVertices(2 * (divisions + 1));
         builder.ensureRectangleIndices(divisions);
